refactor login/signup to dynamic page, mobile style
