+++ client/src/components/Auth/Login.js
@@ -1,0 +1,38 @@
+import React from "react";
+import { Box, Button, FormControl, TextField } from "@material-ui/core";
+
+export const Login = ({ classes }) => {
+  return (
+    <Box>
+      <Box>
+        <FormControl className={classes.formField} margin="normal" required>
+          <TextField
+            aria-label="e-mail address"
+            label="E-mail address"
+            name="username"
+            type="text"
+          />
+        </FormControl>
+      </Box>
+      <FormControl className={classes.formField} margin="normal" required>
+        <TextField
+          label="Password"
+          aria-label="password"
+          type="password"
+          name="password"
+        />
+      </FormControl>
+      <Box className={classes.submitBtnContainer}>
+        <Button
+          className={classes.submitBtn}
+          style={{ fontSize: "1.3rem" }}
+          type="submit"
+          variant="contained"
+          size="large"
+        >
+          Login
+        </Button>
+      </Box>
+    </Box>
+  );
+};
