+++ client/src/routes.js
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { Route, Switch, withRouter } from "react-router-dom";
 import { connect } from "react-redux";
 import { fetchUser } from "./store/utils/thunkCreators";
-import Signup from "./Signup.js";
-import Login from "./Login.js";
+// import Signup from "./Signup.js";
+import Auth from "./Auth.js";
 import { Home, SnackbarError } from "./components";
 
 const Routes = (props) => {
@@ -41,12 +41,11 @@ const Routes = (props) => {
         />
       )}
       <Switch>
-        <Route path="/login" component={Login} />
-        <Route path="/register" component={Signup} />
+        <Route path="/auth" component={Auth} />
         <Route
           exact
           path="/"
-          render={(props) => (props.user?.id ? <Home /> : <Signup />)}
+          render={(props) => (props.user?.id ? <Home /> : <Auth />)}
         />
         <Route path="/home" component={Home} />
       </Switch>
